staging: reverted changes and updated profile page to use new prisma client with SSR
