+++ src/pages/artist/[artist_id]/index.tsx
@@ -1,10 +1,11 @@
+import React from "react";
 import { CircularProgress } from "@mui/material";
 import { useRouter } from "next/router";
-import React from "react";
 import Layout from "../../../components/Layout";
 import ArtistPage from "../../../components/Profile/Artist";
 import { useArtist } from "../../../utils/hooks/useQueryData";
 import { GetServerSideProps } from "next";
+import prisma from "../../../../lib/prisma";
 
 const LoadingIndicator = () => (
   <div style={{ textAlign: "center", padding: "100px 24px" }}>
@@ -12,7 +13,9 @@ const LoadingIndicator = () => (
   </div>
 );
 
-const Artist = () => {
+const Artist = (props) => {
+  console.log(props);
+
   const router = useRouter();
   const artistId = router.query?.artist_id as string;
 
@@ -26,21 +29,20 @@ const Artist = () => {
 };
 
 /** TODO!! check back later if Prisma supports this in serverless environment. */
+export const getServerSideProps: GetServerSideProps = async ({ params }) => {
+  const id = params.artist_id as string;
 
-// export const getServerSideProps: GetServerSideProps = async ({ params }) => {
-//   const id = params.artist_id as string;
+  const artist = await prisma.user.findUnique({
+    where: { id },
+  });
 
-//   const artist = await prisma.user.findUnique({
-//     where: { id },
-//   });
-
-//   return {
-//     props: {
-//       data: {
-//         artist: JSON.parse(JSON.stringify(artist)),
-//       },
-//     },
-//   };
-// };
+  return {
+    props: {
+      data: {
+        artist: JSON.parse(JSON.stringify(artist)),
+      },
+    },
+  };
+};
 
 export default Artist;
